+++ fltk-sys/cfltk/cfl_dialog.h
@@ -169,6 +169,8 @@ char *Fl_dir_chooser(const char *message, const char *fname, int relative);
 
 char *Fl_file_chooser(const char *message, const char *pat, const char *fname, int relative);
 
+int Fl_color_chooser(const char *name, unsigned char *r, unsigned char *g, unsigned char *b, int cmode);
+
 #ifdef __cplusplus
 }
 #endif