+++ src/dialog.rs
@@ -826,3 +826,19 @@ pub fn file_chooser(message: &str, pattern: &str, dir: &str, relative: bool) -> 
         }
     }
 }
+
+/// Spawns a color_chooser dialog. `cmode`: Optional mode for color chooser. Default is -1 if rgb mode.
+pub fn color_chooser(name: &str, cmode: i32) -> Option<(u8, u8, u8)> {
+    unsafe {
+        let name = CString::new(name).unwrap();
+        let mut r = 0;
+        let mut g = 0;
+        let mut b = 0;
+        let ret = Fl_color_chooser(name.as_ptr(), &mut r, &mut g, &mut b, cmode);
+        if ret == 0 {
+            None
+        } else {
+            Some((r, g, b))
+        }
+    }
+}